Add patient

Add patient to visit-based simulation pathway

Usage

add_patient(
  in_system,
  id = parent.frame()$id,
  npat = parent.frame()$npat,
  req_visits = parent.frame()$req_visits,
  patients = parent.frame()$patients,
  resources = parent.frame()$resources,
  t = parent.frame()$t,
  z = parent.frame()$z,
  srv_dist_visit = parent.frame()$srv_dist_visit,
  srv_params_visit = parent.frame()$srv_params_visit,
  mean_los_visit = parent.frame()$mean_los_visit,
  sd_los_visit = parent.frame()$sd_los_visit,
  isr = parent.frame()$isr,
  sd_isr = parent.frame()$sd_isr,
  n_slots = parent.frame()$n_slots,
  end_sr = parent.frame()$end_sr,
  sd_esr = parent.frame()$sd_esr
)

Arguments

in_system

boolean (TRUE or FALSE) - is the patient already in the system? If so, will produce adjusted length of stay and visit rates.

id

integer - ID number

npat

integer - number of patients

req_visits

list - to store require visit vectors for each patient

patients

dataframe - to store information about each patient (from create_patient_df())

resources

dataframe - to capture information on available resources

t

tbc

z

tbc

srv_dist_visit

tbc

srv_params_visit

tbc

mean_los_visit

tbc

sd_los_visit

tbc

isr

Integer vector - initial service rate (ISR) or initial visit rate (IVR) - used in estimation of initial number of visits required for each patient, e.g. c(4, 4)

sd_isr

tbc

n_slots

tbc

end_sr

Integer vector - end service rate (end_sr) or final visit rate (FVR) - used in estimation of final number of visits required for each patient, e.g. c(1, 1)

sd_esr

tbc

Value

List with updated id, npat, req_visits, patients, resources objects

Details

When run add_patient(), it will increment the ID and n_pat, find length of stay, initial visit rate and end visit rate. It will save these results in a dataframe, and then check availability in service and modify resources accordingly.